+++ client/pages/index.js
@@ -3,14 +3,13 @@ import Image from 'next/image'
 import { Inter } from 'next/font/google'
 import styles from '@/styles/Home.module.css'
 
-import { AuthProvider } from '@/context/authContext'
 
 const inter = Inter({ subsets: ['latin'] })
 
 export default function Home() {
   return (
     <>
-    <AuthProvider> 
+  
       <Head>
         <title>Create Next App</title>
         <meta name="description" content="Generated by create next app" />
@@ -112,7 +111,7 @@ export default function Home() {
           </a>
         </div>
       </main>
-      </AuthProvider>
+    
     </>
   )
 }
