+++ client/pages/loginPage.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import Login from '@/components/user/login'
+
+export default function loginPage() {
+  return (
+    <>
+        <Login/>
+    </>
+  )
+}
